Removed the feature related extents tests (NSG 19B) from NSG_DataContentsTests as these will be contained in the OGC features tests. Moved the tile related extents tests (NSG 19B)  from NSG_DataContentsTests into the NSG_TileTests because this will allow them to be skipped if this no tile data. At this point, there is nothing left in NSG_DataContentsTests, so it is removed. SpatialReferenceSystemsTests:    There seem to be several issues here and being unable to address them all, I've moved some of the issues into the LOG.    Firstly, these tests fail when there is no gpkg_tile_matrix table, which is often the case in a features-only geopackage. So I added a test to first ensure the table is present before running the test.    Next, the call to parseWKT is triggering an exception that is not a FactoryException but rather an exception due to a casting problem. I added a catch section to record this exception to the LOG rather than trigger a failure so that the rest of the tests can be run.  Comments have been placed into the code where this casting exception is triggered.  In addition to this exception catch, added additional catch segments to identify additional errors. These changes address issues: #10 – fix #12 (closed) – this is still an issue, these code changes include comments and exception catch code and log entries to try and help identify the issue.
